New arch

+++ app/lib/envfile.server.ts
@@ -1,22 +1,11 @@
 import env from "./env.server"
 import fs from 'fs/promises'
 import path from 'path'
-import { emitter } from "./emitter.server"
+import fileExists from "./fileExists"
 
-export async function getComposeFiles() {
-  const configFolderENV = await readEnvFile()
-  const separator = configFolderENV.COMPOSE_PATH_SEPARATOR || ':'
-  const composeFiles = (configFolderENV.COMPOSE_FILE || 'docker-compose.yml').split(separator)
-  return composeFiles
-}
-
-export async function readEnvFile() {
-  const text = await fs.readFile(path.join(env.configFolder, '.env'), {
-    encoding: 'utf8',
-    flag: 'a+',
-  })
+export function parseEnvFileText(envText: string) {
   const envVars = {} as Record<string, string>
-  text.split('\n').forEach((line) => {
+  envText.split('\n').forEach((line) => {
     const [key, value] = line.split('=')
     if (key && value) {
       envVars[key] = value
@@ -25,40 +14,68 @@ export async function readEnvFile() {
   return envVars
 }
 
-export async function addToDotEnv(filename: string) {
-  const configFolderENV = await readEnvFile()
-  const separator = configFolderENV.COMPOSE_PATH_SEPARATOR || ':'
+export function envToText(envVars: Record<string, string>) {
+  return Object.entries(envVars)
+    .map(([key, value]) => `${key}=${value}`)
+    .join('\n')
+}
+
+export async function readEnvFile(projectFolder: string) {
+  const text = await fs.readFile(
+    path.join(env.configFolder, projectFolder, '.env'),
+    { encoding: 'utf8' }
+  )
+  return parseEnvFileText(text)
+}
+
+const COMPOSE_DEFAULT_SEPARATOR = ':' as const
+
+export async function addToDotEnv(projectFolder: string, filename: string) {
+  const envFile = path.join(env.configFolder, projectFolder, '.env')
+  const envExists = await fileExists(envFile)
+  if (!envExists) {
+    return
+  }
+
+  const configFolderENV = await readEnvFile(projectFolder)
+  const separator = configFolderENV.COMPOSE_PATH_SEPARATOR || COMPOSE_DEFAULT_SEPARATOR
   const fileListText = configFolderENV.COMPOSE_FILE || ''
   const composeFiles = new Set(fileListText.split(separator))
   composeFiles.add(filename)
 
-  const newDotEnv = Object.entries({
+  const newDotEnv = envToText({
     ...configFolderENV,
     COMPOSE_PATH_SEPARATOR: separator,
     COMPOSE_FILE: Array.from(composeFiles).join(separator),
   })
-    .map(([key, value]) => `${key}=${value}`)
-    .join('\n')
 
-  await fs.writeFile(path.join(env.configFolder, '.env'), newDotEnv)
-  emitter.emit('message', `Added ${filename} to .env`)
+  await fs.writeFile(
+    path.join(env.configFolder, projectFolder, '.env'),
+    newDotEnv
+  )
 }
 
-export async function removeFromDotEnv(filename: string) {
-  const configFolderENV = await readEnvFile()
-  const separator = configFolderENV.COMPOSE_PATH_SEPARATOR || ':'
+export async function removeFromDotEnv(projectFolder: string, filename: string) {
+  const envFile = path.join(env.configFolder, projectFolder, '.env')
+  const envExists = await fileExists(envFile)
+  if (!envExists) {
+    return
+  }
+
+  const configFolderENV = await readEnvFile(projectFolder)
+  const separator = configFolderENV.COMPOSE_PATH_SEPARATOR || COMPOSE_DEFAULT_SEPARATOR
   const fileListText = configFolderENV.COMPOSE_FILE || ''
   const composeFiles = new Set(fileListText.split(separator))
   composeFiles.delete(filename)
 
-  const newDotEnv = Object.entries({
+  const newDotEnv = envToText({
     ...configFolderENV,
     COMPOSE_PATH_SEPARATOR: separator,
     COMPOSE_FILE: Array.from(composeFiles).join(separator),
   })
-    .map(([key, value]) => `${key}=${value}`)
-    .join('\n')
 
-  await fs.writeFile(path.join(env.configFolder, '.env'), newDotEnv)
-  emitter.emit('message', `Removed ${filename} from .env`)
+  await fs.writeFile(
+    path.join(env.configFolder, projectFolder, '.env'),
+    newDotEnv
+  )
 }
